refactor: split ssg html modules
