+++ crates/ox_content_ssg/src/html.rs
@@ -1,348 +1,21 @@
 //! HTML page generation for SSG.
 
-use std::fmt::Write as _;
-
 use askama::Template;
 use memchr::memmem;
-use serde::{Deserialize, Serialize};
 
-// =============================================================================
-// Theme Configuration Types
-// =============================================================================
+mod nav;
+mod page;
+mod theme;
 
-/// Theme color configuration.
-#[derive(Debug, Clone, Serialize, Deserialize, Default)]
-pub struct ThemeColors {
-    /// Primary accent color.
-    pub primary: Option<String>,
-    /// Primary color on hover.
-    pub primary_hover: Option<String>,
-    /// Background color.
-    pub background: Option<String>,
-    /// Alternative background color.
-    pub background_alt: Option<String>,
-    /// Main text color.
-    pub text: Option<String>,
-    /// Muted text color.
-    pub text_muted: Option<String>,
-    /// Border color.
-    pub border: Option<String>,
-    /// Code block background color.
-    pub code_background: Option<String>,
-    /// Code block text color.
-    pub code_text: Option<String>,
-}
-
-/// Theme layout configuration.
-#[derive(Debug, Clone, Serialize, Deserialize, Default)]
-pub struct ThemeLayout {
-    /// Sidebar width (CSS value).
-    pub sidebar_width: Option<String>,
-    /// Header height (CSS value).
-    pub header_height: Option<String>,
-    /// Maximum content width (CSS value).
-    pub max_content_width: Option<String>,
-}
-
-/// Theme font configuration.
-#[derive(Debug, Clone, Serialize, Deserialize, Default)]
-pub struct ThemeFonts {
-    /// Sans-serif font stack.
-    pub sans: Option<String>,
-    /// Monospace font stack.
-    pub mono: Option<String>,
-}
-
-/// Theme configuration for entry pages.
-#[derive(Debug, Clone, Serialize, Deserialize, Default)]
-pub struct ThemeEntryPage {
-    /// Landing page presentation mode.
-    pub mode: Option<String>,
-}
-
-/// Theme header configuration.
-#[derive(Debug, Clone, Serialize, Deserialize, Default)]
-pub struct ThemeHeader {
-    /// Logo image URL.
-    pub logo: Option<String>,
-    /// Light mode logo image URL.
-    #[serde(rename = "logoLight")]
-    pub logo_light: Option<String>,
-    /// Dark mode logo image URL.
-    #[serde(rename = "logoDark")]
-    pub logo_dark: Option<String>,
-    /// Whether to render the site name text next to the logo.
-    #[serde(rename = "showSiteNameText")]
-    pub show_site_name_text: Option<bool>,
-    /// Logo width in pixels.
-    pub logo_width: Option<u32>,
-    /// Logo height in pixels.
-    pub logo_height: Option<u32>,
-}
-
-/// Theme footer configuration.
-#[derive(Debug, Clone, Serialize, Deserialize, Default)]
-pub struct ThemeFooter {
-    /// Footer message (supports HTML).
-    pub message: Option<String>,
-    /// Copyright text (supports HTML).
-    pub copyright: Option<String>,
-}
-
-/// Social links configuration.
-#[derive(Debug, Clone, Serialize, Deserialize, Default)]
-pub struct SocialLinks {
-    /// GitHub URL.
-    pub github: Option<String>,
-    /// Twitter/X URL.
-    pub twitter: Option<String>,
-    /// Discord URL.
-    pub discord: Option<String>,
-    /// Custom social links.
-    pub links: Option<Vec<SocialLink>>,
-}
-
-/// Custom social link.
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct SocialLink {
-    /// Icon label or built HTML icon source.
-    pub icon: Option<String>,
-    /// Trusted inline SVG icon.
-    pub icon_svg: Option<String>,
-    /// Link URL.
-    pub link: String,
-    /// Accessible label.
-    pub aria_label: Option<String>,
-}
-
-/// Embedded HTML content for specific positions in the page layout.
-#[derive(Debug, Clone, Serialize, Deserialize, Default)]
-pub struct ThemeEmbed {
-    /// Content to inject into `<head>`.
-    pub head: Option<String>,
-    /// Content before header.
-    pub header_before: Option<String>,
-    /// Content after header.
-    pub header_after: Option<String>,
-    /// Content before sidebar navigation.
-    pub sidebar_before: Option<String>,
-    /// Content after sidebar navigation.
-    pub sidebar_after: Option<String>,
-    /// Content before main content.
-    pub content_before: Option<String>,
-    /// Content after main content.
-    pub content_after: Option<String>,
-    /// Content before footer.
-    pub footer_before: Option<String>,
-    /// Custom footer content (replaces default footer).
-    pub footer: Option<String>,
-}
-
-/// Complete theme configuration.
-#[derive(Debug, Clone, Serialize, Deserialize, Default)]
-pub struct ThemeConfig {
-    /// Light mode colors.
-    pub colors: Option<ThemeColors>,
-    /// Dark mode colors.
-    pub dark_colors: Option<ThemeColors>,
-    /// Font configuration.
-    pub fonts: Option<ThemeFonts>,
-    /// Entry page configuration.
-    #[serde(rename = "entryPage")]
-    pub entry_page: Option<ThemeEntryPage>,
-    /// Layout configuration.
-    pub layout: Option<ThemeLayout>,
-    /// Header configuration.
-    pub header: Option<ThemeHeader>,
-    /// Footer configuration.
-    pub footer: Option<ThemeFooter>,
-    /// Social links configuration.
-    pub social_links: Option<SocialLinks>,
-    /// Embedded HTML content at specific positions.
-    pub embed: Option<ThemeEmbed>,
-    /// Additional custom CSS.
-    pub css: Option<String>,
-    /// Additional custom JavaScript.
-    pub js: Option<String>,
-}
-
-// =============================================================================
-// Entry Page Types
-// =============================================================================
-
-/// Hero action button.
-#[derive(Debug, Clone, Serialize, Deserialize, Default)]
-pub struct HeroAction {
-    /// Button theme: "brand" or "alt".
-    pub theme: Option<String>,
-    /// Button text.
-    pub text: String,
-    /// Link URL.
-    pub link: String,
-}
-
-/// Hero image configuration.
-#[derive(Debug, Clone, Serialize, Deserialize, Default)]
-pub struct HeroImage {
-    /// Image source URL.
-    pub src: String,
-    /// Light mode image source URL.
-    #[serde(rename = "lightSrc")]
-    pub light_src: Option<String>,
-    /// Dark mode image source URL.
-    #[serde(rename = "darkSrc")]
-    pub dark_src: Option<String>,
-    /// Alt text.
-    pub alt: Option<String>,
-    /// Image width.
-    pub width: Option<u32>,
-    /// Image height.
-    pub height: Option<u32>,
-}
-
-/// Hero section configuration.
-#[derive(Debug, Clone, Serialize, Deserialize, Default)]
-pub struct HeroConfig {
-    /// Main title (large, gradient text).
-    pub name: Option<String>,
-    /// Secondary text.
-    pub text: Option<String>,
-    /// Tagline.
-    pub tagline: Option<String>,
-    /// Optional notice shown in the hero.
-    pub notice: Option<HeroNoticeConfig>,
-    /// Hero image.
-    pub image: Option<HeroImage>,
-    /// Action buttons.
-    pub actions: Option<Vec<HeroAction>>,
-}
-
-/// Hero notice configuration.
-#[derive(Debug, Clone, Serialize, Deserialize, Default)]
-pub struct HeroNoticeConfig {
-    /// Notice title.
-    pub title: Option<String>,
-    /// Notice paragraphs.
-    pub body: Option<Vec<String>>,
-}
-
-/// Feature card configuration.
-#[derive(Debug, Clone, Serialize, Deserialize, Default)]
-pub struct FeatureConfig {
-    /// Icon - supports: "mdi:icon-name" (Iconify), image URL, or emoji.
-    pub icon: Option<String>,
-    /// Feature title.
-    pub title: String,
-    /// Feature description.
-    pub details: Option<String>,
-    /// Optional link.
-    pub link: Option<String>,
-    /// Link text.
-    pub link_text: Option<String>,
-}
-
-/// Entry page configuration (for landing pages with hero and features).
-#[derive(Debug, Clone, Serialize, Deserialize, Default)]
-pub struct EntryPageConfig {
-    /// Hero section.
-    pub hero: Option<HeroConfig>,
-    /// Feature cards.
-    pub features: Option<Vec<FeatureConfig>>,
-}
-
-// =============================================================================
-// Navigation and Page Types
-// =============================================================================
-
-/// Navigation item for SSG.
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct NavItem {
-    /// Display title.
-    pub title: String,
-    /// URL path.
-    pub path: String,
-    /// Full href.
-    pub href: String,
-    #[serde(default)]
-    pub children: Vec<NavItem>,
-    #[serde(default)]
-    pub collapsed: Option<bool>,
-    #[serde(default)]
-    pub sticky_collapsed: Option<bool>,
-}
-
-/// Navigation group for SSG.
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct NavGroup {
-    /// Group title.
-    pub title: String,
-    /// Navigation items.
-    pub items: Vec<NavItem>,
-    #[serde(default)]
-    pub collapsed: Option<bool>,
-    #[serde(default)]
-    pub sticky_collapsed: Option<bool>,
-}
-
-/// Table of contents entry.
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct TocEntry {
-    /// Heading depth (1-6).
-    pub depth: u8,
-    /// Heading text.
-    pub text: String,
-    /// URL-friendly slug.
-    pub slug: String,
-}
-
-/// Page data for SSG.
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct PageData {
-    /// Page title.
-    pub title: String,
-    /// Page description.
-    pub description: Option<String>,
-    /// Page content HTML.
-    pub content: String,
-    /// Table of contents entries.
-    pub toc: Vec<TocEntry>,
-    /// Last updated timestamp in milliseconds since the Unix epoch.
-    pub last_updated: Option<i64>,
-    /// URL path.
-    pub path: String,
-    /// Entry page configuration (if layout: entry).
-    pub entry_page: Option<EntryPageConfig>,
-}
-
-/// SSG configuration.
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct SsgConfig {
-    /// Site name.
-    pub site_name: String,
-    /// Base URL path.
-    pub base: String,
-    /// OG image URL.
-    pub og_image: Option<String>,
-    /// Theme configuration.
-    pub theme: Option<ThemeConfig>,
-    /// Current locale (BCP 47 tag) for this page, if i18n is enabled.
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub locale: Option<String>,
-    /// All available locales (for generating locale switcher and hreflang tags).
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub available_locales: Option<Vec<LocaleInfo>>,
-}
-
-/// Locale information for the locale switcher.
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct LocaleInfo {
-    /// BCP 47 locale tag.
-    pub code: String,
-    /// Display name.
-    pub name: String,
-    /// Text direction.
-    pub dir: String,
-}
+use nav::generate_nav_html;
+pub use page::{
+    EntryPageConfig, FeatureConfig, HeroAction, HeroConfig, HeroImage, HeroNoticeConfig,
+    LocaleInfo, NavGroup, NavItem, PageData, SsgConfig, TocEntry,
+};
+pub use theme::{
+    SocialLink, SocialLinks, ThemeColors, ThemeConfig, ThemeEmbed, ThemeEntryPage, ThemeFonts,
+    ThemeFooter, ThemeHeader, ThemeLayout,
+};
 
 // =============================================================================
 // Askama Template Structures
@@ -1190,108 +863,6 @@ fn validate_social_svg(svg: &str) -> Option<&str> {
         Some(trimmed)
     } else {
         None
-    }
-}
-
-fn generate_nav_html(nav_groups: &[NavGroup], current_path: &str) -> String {
-    let mut html = String::new();
-    for (group_index, group) in nav_groups.iter().enumerate() {
-        if group.collapsed.is_some() {
-            let open = if group.collapsed == Some(true) { "" } else { " open" };
-            let state_key = format!("group:{group_index}:{}", group.title);
-            let state_attr = nav_state_attr(group.sticky_collapsed, &state_key);
-            push_fmt(&mut html, format_args!(
-                "<details class=\"nav-section nav-section--collapsible\"{open}{state_attr}>\n  <summary class=\"nav-title nav-title--summary\">{}</summary>\n",
-                escape_html(&group.title)
-            ));
-            render_nav_list(
-                &mut html,
-                &group.items,
-                current_path,
-                false,
-                &format!("group:{group_index}"),
-            );
-            html.push_str("</details>\n");
-        } else {
-            push_fmt(
-                &mut html,
-                format_args!(
-                    "<div class=\"nav-section\">\n  <div class=\"nav-title\">{}</div>\n",
-                    escape_html(&group.title)
-                ),
-            );
-            render_nav_list(
-                &mut html,
-                &group.items,
-                current_path,
-                false,
-                &format!("group:{group_index}"),
-            );
-            html.push_str("</div>\n");
-        }
-    }
-    html
-}
-
-fn render_nav_list(
-    html: &mut String,
-    items: &[NavItem],
-    current_path: &str,
-    nested: bool,
-    key_prefix: &str,
-) {
-    let class_name = if nested { "nav-list nav-list--nested" } else { "nav-list" };
-    push_fmt(html, format_args!("  <ul class=\"{class_name}\">\n"));
-    for (index, item) in items.iter().enumerate() {
-        render_nav_item(html, item, current_path, &format!("{key_prefix}.{index}"));
-    }
-    html.push_str("  </ul>\n");
-}
-
-fn render_nav_item(html: &mut String, item: &NavItem, current_path: &str, key_path: &str) {
-    let href = safe_nav_href(&item.href);
-    let title = escape_html(&item.title);
-    let active_class = if item.path == current_path { " active" } else { "" };
-    if item.children.is_empty() {
-        push_fmt(html, format_args!(
-            "    <li class=\"nav-item\"><a href=\"{href}\" class=\"nav-link{active_class}\">{title}</a></li>\n"
-        ));
-        return;
-    }
-
-    let open = if item.collapsed == Some(true) { "" } else { " open" };
-    let state_key = format!("item:{key_path}:{}:{}", item.path, item.title);
-    let state_attr = nav_state_attr(item.sticky_collapsed, &state_key);
-    push_fmt(html, format_args!(
-        "    <li class=\"nav-item nav-item--group\"><details class=\"nav-details\"{open}{state_attr}><summary class=\"nav-summary\"><a href=\"{href}\" class=\"nav-link nav-link--summary{active_class}\">{title}</a></summary>\n"
-    ));
-    render_nav_list(html, &item.children, current_path, true, key_path);
-    html.push_str("    </details></li>\n");
-}
-
-fn nav_state_attr(sticky_collapsed: Option<bool>, key: &str) -> String {
-    if sticky_collapsed == Some(true) {
-        format!(" data-ox-nav-state-key=\"{}\"", escape_html(key))
-    } else {
-        String::new()
-    }
-}
-
-fn safe_nav_href(href: &str) -> String {
-    let trimmed = href.trim();
-    let lower = trimmed.to_ascii_lowercase();
-    let safe_scheme = lower.starts_with("http://")
-        || lower.starts_with("https://")
-        || lower.starts_with("mailto:");
-    if trimmed.starts_with("//") || (trimmed.contains(':') && !safe_scheme) {
-        return "#".to_string();
-    }
-    escape_html(trimmed)
-}
-
-fn push_fmt(output: &mut String, args: std::fmt::Arguments<'_>) {
-    if output.write_fmt(args).is_err() {
-        output.push_str("[formatting failed]");
     }
 }
